requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.9
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
retrying==1.3.4
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
sklearn-compat==0.1.3
tenacity==9.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
Werkzeug==3.0.6
zipp==3.21.0
